requirements.txt
faiss-cpu
numpy
pandas
langchain
langchain-community
langchain-openai
sentence-transformers
tqdm
transformers
langchain-cpp